refactor : comment 전략 패턴 수정
- 댓글 유효성 검사시 전략 패턴 적용했었음
- 추후 전략 확장을 쉽게 하기 위해서 전략 컨텍스트 추가
