+++ comment-service/src/main/java/com/sparta/moim/comment/domain/strategy/validation/CommentValidation.java
@@ -1,13 +1,10 @@
-package com.sparta.moim.comment.domain.validation;
+package com.sparta.moim.comment.domain.strategy.validation;
 
 import static com.sparta.moim.comment.infrastructure.response.CommentCode.PARENT_COMMENT_CANNOT_HAVE_COMMENT;
-import static com.sparta.moim.comment.infrastructure.response.CommentCode.PARENT_COMMENT_NOT_FOUND;
-import static com.sparta.moim.comment.infrastructure.response.CommentCode.POST_NOT_FOUND;
 
 import com.sparta.moim.comment.domain.repository.CommentRepository;
 import com.sparta.moim.comment.presentation.request.CommentRequestDTO;
 import com.sparta.moim.common.exception.BaseException;
-import java.util.UUID;
 import org.springframework.stereotype.Component;
 
 @Component
